frostedlib:dimension_shift additions and fixes.
Fixed some of the liquid handling such that platforms now spawn at the surface of the liquid and place the player at the surface on the platform instead of at the bottom of the liquid.

Added a new variable, strict_height, with it set to true, it will require that the height follows the target_height. Meaning it will not use a height that doesn't follow it as a fallback.